# Use section.totalSize to figure out item count in section