520cf77 removed `@memoize`, so the `all()` was actually called twice.

This instead makes a simple API call with no data:

```
"GET /library/sections/1/all?includeCollections=0&X-Plex-Container-Size=0&X-Plex-Container-Start=0 HTTP/1.1" 200 335
```

For my library of 644 items (cache is filled) the timing is:
- before: 9.7s, 10.0s, 9.1s
- after: 6.6s, 6.4s, 6.4s